Splitted main logic to be more readable
